added url_parser module and updated tests accordingly
