+++ tests/conftest.py
@@ -9,8 +9,9 @@ import pytest
 import valkey
 from packaging.version import Version
 from valkey import Sentinel
+from valkey._parsers import parse_url
 from valkey.backoff import NoBackoff
-from valkey.connection import Connection, parse_url
+from valkey.connection import Connection
 from valkey.exceptions import ValkeyClusterException
 from valkey.retry import Retry
 
@@ -275,7 +276,7 @@ def _get_client(
 
     cluster_mode = VALKEY_INFO["cluster_enabled"]
     if not cluster_mode:
-        url_options = parse_url(valkey_url)
+        url_options = parse_url(valkey_url, False)
         url_options.update(kwargs)
         pool = valkey.ConnectionPool(**url_options)
         client = cls(connection_pool=pool)
