+++ valkey/cluster.py
@@ -6,13 +6,13 @@ import time
 from collections import OrderedDict
 from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 
-from valkey._parsers import CommandsParser, Encoder
+from valkey._parsers import CommandsParser, Encoder, parse_url
 from valkey._parsers.helpers import parse_scan
 from valkey.backoff import default_backoff
 from valkey.client import CaseInsensitiveDict, PubSub, Valkey
 from valkey.commands import READ_COMMANDS, ValkeyClusterCommands
 from valkey.commands.helpers import list_or_args
-from valkey.connection import ConnectionPool, DefaultParser, parse_url
+from valkey.connection import ConnectionPool, DefaultParser
 from valkey.crc import VALKEY_CLUSTER_HASH_SLOTS, key_slot
 from valkey.exceptions import (
     AskError,
@@ -580,7 +580,7 @@ class ValkeyCluster(AbstractValkeyCluster, ValkeyClusterCommands):
         from_url = False
         if url is not None:
             from_url = True
-            url_options = parse_url(url)
+            url_options = parse_url(url, False)
             if "path" in url_options:
                 raise ValkeyClusterException(
                     "ValkeyCluster does not currently support Unix Domain "
